basic point decoding for new garmin maps
git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@1670 d0c010d8-bb82-477e-a075-112140230451

+++ src/CMapTDB.h
@@ -305,7 +305,7 @@ class CMapTDB : public IMap
         QMap<quint32, polygon_property> polygonProperties;
         QList<quint16> polygonDrawOrder;
 
-        QMap<quint16, QImage> pointProperties;
+        QMap<quint32, QImage> pointProperties;
 
         polytype_t polygons;
         polytype_t polylines;
